Hoist the composition types to the root and restructure queries
