+++ persistence/source/contents/all/index.ts
@@ -2,7 +2,7 @@ import path from "path";
 import { PostContents } from "@domain/post";
 import { onlyMdx } from "@persistence/utils";
 import { PROJECTS_DIRECTORY, BLOG_DIRECTORY } from "@persistence/const";
-import { Dependencies, di, QueryKind } from "../composition";
+import { Dependencies, di, QueryKind } from "@persistence/composition";
 
 type Executor = (di?: Dependencies) => List<PostContents>;
