+++ persistence/source/contents/one/index.ts
@@ -2,7 +2,7 @@ import path from "path";
 import { PostContents, PostId } from "@domain/post";
 import { withMdx } from "@persistence/utils";
 import { PROJECTS_DIRECTORY, BLOG_DIRECTORY } from "@persistence/const";
-import { Dependencies, di, QueryKind } from "../composition";
+import { Dependencies, di, QueryKind } from "@persistence/composition";
 
 type Executor = (id: PostId, di?: Dependencies) => PostContents;
